+++ Programs/swap.java
@@ -1,0 +1,13 @@
+class swap
+{
+      //Program to swap
+      public static void main(String[ ]args)
+      {
+            int a=5,b=10,c;
+            System.out.println("NUMBERS BEFORE SWAPPING "+a+ ","+b);
+            c=a;
+            a=b;
+            b=c;
+            System.out.println("NUMBERS AFTER SWAPPING "+a+ ","+b);
+       }
+}
