--- name.java
@@ -1,8 +1,0 @@
-class name
-{
-      //Program to display a message
-      public static void main(String[ ]args)
-      {
-            System.out.println("MY NAME IS "+ " "+args[0]+ " "+args[1]+ " "+args[2]);
-       }
-}
