--- sum.java
@@ -1,10 +1,0 @@
-class sum
-{
-      //Program to find the sum
-      public static void main(String[ ]args)
-      {
-            int a=5,b=10,sum;
-            sum=a+b;
-            System.out.println("SUM IS "+sum);
-       }
-}
